tg_bot_files/requirements.txt
aiogram==3.10.0
apscheduler
python-dotenv
